This whole writeBuffers[] business was never implemented properly.
Moreover I don't think it should ever be possible for it to happen. We
buffer an unlimited amount in the tcp pipe so we should always be able
to write the whole thing.

Additionally bytesRead and bytesWritten should be logical bytes, not
encrypted bytes (I think, it seems curl does that).